backend/requirements.txt
crewai
openai
python-dotenv
requests
pillow
moviepy